Make template Hotwire ready
